get example working
